KAFKA-9492; Ignore record errors in ProduceResponse for older versions (#8030)
Fixes NPE in brokers when processing record errors in produce response for older versions.

Reviewers: Chia-Ping Tsai <chia7712@gmail.com>, Ismael Juma <ismael@juma.me.uk>, Jason Gustafson <jason@confluent.io>, Guozhang Wang <wangguoz@gmail.com>
